Code changes to help compile in MSVC 2010. Updated cmake build files to use latest sources. Added cmake stanza to build and use libtiff 4.0.1, if it's there, on Windows. Added cmake stanza to use boost files under src_packages, if it's unpacked already.

+++ v3d_main/neuron_annotator/gui/DynamicRangeTool.h
@@ -35,7 +35,7 @@ protected:
 
 private:
     Ui::DynamicRangeTool ui;
-    size_t currentChannelIndex;
+    int currentChannelIndex;
     DataColorModel* dataColorModel;
     QColor channelColor;
     int channelHdrMin;
